+++ HISTORY.md
@@ -13,6 +13,9 @@ Our backwards-compatibility policy can be found [here](https://github.com/python
 
 ## 25.4.0 (UNRELEASED)
 
+- Add the {mod}`tomllib <cattrs.preconf.tomllib>` preconf converter.
+  See [here](https://catt.rs/en/latest/preconf.html#tomllib) for details.
+  ([#716](https://github.com/python-attrs/cattrs/pull/716))
 - Fix structuring of nested generic classes with stringified annotations.
   ([#688](https://github.com/python-attrs/cattrs/pull/688))
 - Python 3.9 is no longer supported, as it is end-of-life. Use previous versions on this Python version.